+++ public/admin/_app/immutable/nodes/24.XdIHOnx7.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/YWg_BxjS.js";export{e as component};